(Qt) Combine filedropwidgets.cpp and settingswidgets.cpp into qt_widgets.cpp - further binary size reduction
