+++ Makefile.common
@@ -490,14 +490,13 @@ ifeq ($(HAVE_QT), 1)
           ui/drivers/qt/ui_qt_load_core_window.o \
           ui/drivers/qt/gridview.o \
           ui/drivers/qt/qt_dialogs.o \
-          ui/drivers/qt/filedropwidget.o \
+          ui/drivers/qt/qt_widgets.o \
           ui/drivers/qt/qt_playlist.o \
           ui/drivers/qt/thumbnaildownload.o \
           ui/drivers/qt/thumbnailpackdownload.o \
           ui/drivers/qt/playlistthumbnaildownload.o
     ifeq ($(HAVE_MENU), 1)
-       OBJ += ui/drivers/qt/settingswidgets.o \
-              ui/drivers/qt/options/achievements.o \
+       OBJ += ui/drivers/qt/options/achievements.o \
               ui/drivers/qt/options/audio.o \
               ui/drivers/qt/options/generic.o \
               ui/drivers/qt/options/input.o \
