+++ griffin/griffin_cpp.cpp
@@ -54,13 +54,12 @@ UI
 #include "../ui/drivers/qt/ui_qt_load_core_window.cpp"
 #include "../ui/drivers/qt/gridview.cpp"
 #include "../ui/drivers/qt/qt_dialogs.cpp"
-#include "../ui/drivers/qt/filedropwidget.cpp"
+#include "../ui/drivers/qt/qt_widgets.cpp"
 #include "../ui/drivers/qt/qt_playlist.cpp"
 #include "../ui/drivers/qt/thumbnaildownload.cpp"
 #include "../ui/drivers/qt/thumbnailpackdownload.cpp"
 #include "../ui/drivers/qt/playlistthumbnaildownload.cpp"
 #ifdef HAVE_MENU
-#include "../ui/drivers/qt/settingswidgets.cpp"
 #include "../ui/drivers/qt/options/generic.cpp"
 #include "../ui/drivers/qt/options/video.cpp"
 #include "../ui/drivers/qt/options/audio.cpp"
